Working text input
